+++ oscqam/actions.py
@@ -11,6 +11,8 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 import osc.conf
 
 from .errors import (NoCommentError,
+                     NonMatchingGroupsError,
+                     NonMatchingUserGroupsError,
                      NotAssignedError,
                      NotPreviousReviewerError,
                      NoReviewError,
@@ -527,17 +529,59 @@ class UnassignAction(OscAction):
 
 
 class ApproveAction(OscAction):
+    """Template class for Approval actions.
+
+    Subclasses need to overwrite:
+
+    - get_reviewer: whose review is done.
+    """
+
+    __metaclass__ = abc.ABCMeta
+
+    def __init__(self, remote, user, request_id, reviewer,
+                 template_factory = Template, out = sys.stdout):
+        """Approve a review for either a User or a Group.
+
+        :param remote: Remote interface for build service calls.
+        :type remote: L{oscqam.remote.RemoteFacade}
+
+        :param user: The user performing this action.
+        :type user: L{string}
+
+        :param request_id: Id of the request to accept.
+        :type request_id: L{int}
+
+        :param reviewer: Reviewer to accept this request for.
+        :type reviewer: L{oscqam.models.User} | L{oscqam.models.Group}
+
+        :param template_factory: Function to get a report-template from.
+        :type template_factory:
+
+        :param out: File like object to write output messages to.
+        :type out:
+        """
+        super(ApproveAction, self).__init__(remote, user, out)
+        self.request = remote.requests.by_id(request_id)
+        self.template = self.request.get_template(template_factory)
+        self.reviewer = self.get_reviewer(reviewer)
+
+    @abc.abstractmethod
+    def get_reviewer(self, reviwer):
+        """Return the object for the given reviewer."""
+        pass
+
+
+class ApproveUserAction(ApproveAction):
     """Approve a review for a user.
 
     """
     APPROVE_MSG = ("Approving {request} for {user} ({groups}). "
                    "Testreport: {url}")
+    MORE_GROUPS_MSG = ("The following groups could also be reviewed by you: "
+                       "{groups}")
 
-    def __init__(self, remote, user, request_id, template_factory = Template,
-                 out = sys.stdout):
-        super(ApproveAction, self).__init__(remote, user, out = out)
-        self.request = remote.requests.by_id(request_id)
-        self.template = self.request.get_template(template_factory)
+    def get_reviewer(self, reviewer):
+        return self.remote.users.by_name(reviewer)
 
     def reviews_assigned(self):
         """Ensure that the user was assigned before accepting."""
@@ -559,16 +603,45 @@ class ApproveAction(OscAction):
         self.template.testplanreviewer()
         self.template.passed()
 
+    def additional_reviews(self):
+        """Return groups that could also be reviewed by the user."""
+        return self.user.reviewable_groups(self.request)
+
     def action(self):
         self.validate()
         url = self.template.url()
         groups = ", ".join([str(g) for g in self.user.in_review_groups(self.request)])
-        msg = ApproveAction.APPROVE_MSG.format(user = self.user,
-                                               groups = groups,
-                                               request = self.request,
-                                               url = url)
+        msg = self.APPROVE_MSG.format(user = self.reviewer,
+                                      groups = groups,
+                                      request = self.request,
+                                      url = url)
         self.print(msg)
-        self.request.review_accept(user = self.user,
+        self.request.review_accept(user = self.reviewer,
+                                   comment = msg)
+        try:
+            groups = ', '.join([str(g) for g in self.additional_reviews()])
+            msg = self.MORE_GROUPS_MSG.format(groups = groups)
+            self.print(msg)
+        except NonMatchingUserGroupsError:
+            pass
+
+
+class ApproveGroupAction(ApproveAction):
+    APPROVE_MSG = "Approving {request} for group {group}."
+
+    def get_reviewer(self, reviewer):
+        return self.remote.groups.for_name(reviewer)
+
+    def validate(self):
+        if self.reviewer not in self.request.groups:
+            raise NonMatchingGroupsError([self.reviewer], self.request.groups)
+
+    def action(self):
+        self.validate()
+        msg = self.APPROVE_MSG.format(request = self.request,
+                                      group = self.reviewer)
+        self.print(msg)
+        self.request.review_accept(group = self.reviewer,
                                    comment = msg)